Fixed an issue with jQuery.fn.textNodes() not working in IE9. Ticket #2
This issue seems to have bee caused by an issue with IE9's implementation
of Text.splitText() that did not update the parents .childNodes property.
This meant that jQuery.fn.contents() was returning the incorrect results.
jQuery.fn.textNodes() now manually traverses an elements children in order
to work around this issue.