Always specify XPath index explicitly.
This is preparation for dealing effectively with changing content, and
addresses the following problem. Say the following [[annotation]] is made:

    <p>Hi there, this is <em>an [[emphasised]] text</em></p>

This will (currently) be serialized as [('/p/em', 3), ('/p/em', 13)]. If I
modify the document to become

    <p>I broke it.</p>

    <p>Hi there, this is <em>an emphasised text</em></p>

Then the deserializer will actually work, because '/p/em' (with unspecified
indices) is unambiguous in the document. But this is really just an accident,
and is inconsistent with the fact that if the original document were

    <p>Yet another paragraph</p>

    <p>Hi there, this is <em>an [[emphasised]] text</em></p>

the annotation would be serialized as [('/p[2]/em', 3), ('/p[2]/em', 13)] and
an identical transformation of the document, to

    <p>I broke it.</p>

    <p>Yet another paragraph</p>

    <p>Hi there, this is <em>an [[emphasised]] text</em></p>

will now cause the deserializer to break, because '/p[2]/em' doesn't match
anything.

I think these two cases should behave identically, and so $(...).xpath() now
emits explicit indices all the time.