Fix an off-by-one problem with range normalization.
When creating a BrowserRange from a SerializedRange (for normalizing),
and specifying the start (with a node/offset pair),
never use the very end of a text node; always use the start of the next one.

Specifying the end of a text node (where there are no more characters left)
can have unwanted consequences.

(I have added test case to showcase the problem.)