Attempt to clarify range logic with comments
I switched to using a strict greater than comparison here because
it privileges the start offset: rather than adding 1 to the start,
we remove 1 from the end which, to me, is a clearer reflection of
what's going on. The Range specification describes the endOffset as
the boundary offset, therefore excluding the character at that index.