range: Improved test for an off-by-one error
The test removed by this commit was originally introduced in order to
demonstrate a failing case that the addition of [this line][1] fixed in
normalisation of serialized ranges.

[1]: https://github.com/openannotation/xpath-range/blob/b1827dd16b54d23e632c3da5fca4860c8d0dc080/src/range.coffee#L336

This commit replaces that test with a pair of tests that more clearly
demonstrate the intent and purpose of the test.

At the same time, it removes the Util.readRangeViaSelection function
that is only used in this test.