Test NormalizedRange#limit more thoroughly
Previously, only the left-hand side was tested; and grandchildren of the
bounds were not tested at all. With [] indicating a text node:

    <h1>[My Heading]</h1>
    <p>
      [My paragraph]
      <span>
        [ conti]
        [nues]
      </span>
    <p>
    <p>[Another paragraph begins]</p>

I wrote an implemetation where, given a range ending in the second
paragraph, limiting it to the first paragraph would incorrectly update
the `@end` to ` conti` not `nues`. The old test case did not catch it.