Don’t assume start node has a nextSibling
If a range start offset as the very end of its container node, we want
to move the cursor forward to the beginning of the next node in the DOM.
However, the next node is not necessarily a sibling of the start
container reported by the browser, it may be a sibling of its parent,
its parent’s parent, etc. So, be sure to move up the tree until we find
a node that has a next sibling.