Update LICENSE
Use the "Annotator contributors" copyright assignment, as this is the
most accurate thing for the inheritance of Annotator and conveys the
intention to have contributors maintain copyright over their own
contributions. However, as there is no significant code remaining that
has not undergone a recent rewrite, change the license on the code as
it now stands, incorporating any remaining derivation under the terms
of its original licensing as MIT (in addition to GPL).